==> 02-scripts/01-emorie/03-Q2.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Question 2: Bivariate Change As Outcome Models of Personality States and Situation Characteristics  

## Data Setup  
### Idiographic  
#### Study 1: PAIRS  
```{r q2 load pairs}
load(sprintf("%s/03-data/01-pairs/02-clean-data/pairs-clean-edb.RData", wd))
```

```{r q2 pairs idio setup}
pairs_states_long <- pairs_states_long %>%
  mutate(value2 = value^2)

q2_idio_pairs_nested <- pairs_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    pairs_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  nest() %>%
  ungroup()
```

```{r q2 pairs idio save}
save_fun <- function(d, sid, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/01-pairs/05-q2-idio/%s-%s-%s-%s.RData", wd, sid, pname, pitem, sname))
}

q2_idio_pairs_nested %>%
  mutate(d = pmap(list(data, SID, p_name, p_item, s_name), save_fun))
```

#### Study 2: IPCS  
```{r q2 load ipcs}
load(sprintf("%s/03-data/02-ipcs/02-clean-data/ipcs-clean-edb.RData", wd))
```

```{r q2 ipcs idio setup}
ipcs_states_long <- ipcs_states_long %>%
  mutate(value2 = value^2)

q2_idio_ipcs_nested <- ipcs_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    ipcs_states_long %>% 
      filter(category == "sit") %>%
      select(SID, tdif, Date, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  nest() %>%
  ungroup()
```

```{r q2 ipcs idio save}
save_fun <- function(d, sid, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/02-ipcs/05-q2-idio/%s-%s-%s-%s.RData", wd, sid, pname, pitem, sname))
}

q2_idio_ipcs_nested %>%
  mutate(d = pmap(list(data, SID, p_name, p_item, s_name), save_fun))
```

#### Study 3: Sherman et al., 2015  
```{r q2 load sherman}
load(sprintf("%s/03-data/03-sherman/02-clean-data/sherman-clean-edb.RData", wd))
```

```{r q2 sherman idio setup}
sherman_states_long <- sherman_states_long %>%
  mutate(value2 = value^2)

q2_idio_sherman_nested <- sherman_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    sherman_states_long %>% 
      filter(category == "sit") %>%
      select(SID, tdif, Date, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  nest() %>%
  ungroup()
```

```{r q2 sherman idio save}
save_fun <- function(d, sid, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/03-sherman/05-q2-idio/%s-%s-%s-%s.RData", wd, sid, pname, pitem, sname))
}

q2_idio_sherman_nested %>%
  mutate(d = pmap(list(data, SID, p_name, p_item, s_name), save_fun))
```

#### Study 4: Horstmann et al., 2021  

```{r q2 load horstmann}
load(sprintf("%s/03-data/04-horstmann/02-clean-data/horstmann-clean-edb.RData", wd))
```

```{r q2 horstmann idio setup}
horstmann_states_long <- horstmann_states_long %>%
  mutate(value2 = value^2)

q2_idio_horstmann_nested <- horstmann_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    horstmann_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  nest() %>%
  ungroup()
```

```{r q2 horstmann idio save}
save_fun <- function(d, sid, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/04-horstmann/05-q2-idio/%s-%s-%s-%s.RData", wd, sid, pname, pitem, sname))
}

q2_idio_horstmann_nested %>%
  mutate(d = pmap(list(data, SID, p_name, p_item, s_name), save_fun))
```

### Multilevel  
#### Study 1: PAIRS  
```{r q2 load pairs}
load(sprintf("%s/03-data/01-pairs/02-clean-data/pairs-clean-edb.RData", wd))
```

```{r q2 pairs mlm setup}
pairs_states_long <- pairs_states_long %>%
  mutate(value2 = value^2)

q2_mlm_pairs_nested <- pairs_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    pairs_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  mutate(p_value_c = p_value - mean(p_value, na.rm = T)
         , s_value_c = s_value - mean(s_value, na.rm = T)) %>%
  group_by(p_name, p_item, s_name) %>%
  nest() %>%
  ungroup()
```

```{r q2 pairs mlm save}
save_fun <- function(d, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/01-pairs/06-q2-mlm/%s-%s-%s.RData", wd, pname, pitem, sname))
}

q2_mlm_pairs_nested %>%
  mutate(d = pmap(list(data, p_name, p_item, s_name), save_fun))
```

```{r}
length(unique((q2_mlm_pairs_nested %>% unnest(data))$SID))
```


#### Study 2: IPCS  
```{r q2 load ipcs}
load(sprintf("%s/03-data/02-ipcs/02-clean-data/ipcs-clean-edb.RData", wd))
```

```{r q2 ipcs mlm setup}
ipcs_states_long <- ipcs_states_long %>%
  mutate(value2 = value^2)

q2_mlm_ipcs_nested <- ipcs_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    ipcs_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  mutate(p_value_c = p_value - mean(p_value, na.rm = T)
         , s_value_c = s_value - mean(s_value, na.rm = T)) %>%
  group_by(p_name, p_item, s_name) %>%
  nest() %>%
  ungroup()
```

```{r q2 ipcs mlm save}
save_fun <- function(d, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/02-ipcs/06-q2-mlm/%s-%s-%s.RData", wd, pname, pitem, sname))
}

q2_mlm_ipcs_nested %>%
  mutate(d = pmap(list(data, p_name, p_item, s_name), save_fun))
```


```{r}
length(unique((q2_mlm_ipcs_nested %>% unnest(data))$SID))
q2_mlm_ipcs_nested %>% unnest(data) %>%
  select(SID, Date) %>% distinct() %>%
  group_by(SID) %>%
  tally() %>%
  ungroup() %>% 
  mutate(n = n+1) %>% 
  summarize_at(vars(n), lst(min, max, mean, sd))
```


#### Study 3: Sherman et al., 2015  
```{r q2 load sherman}
load(sprintf("%s/03-data/03-sherman/02-clean-data/sherman-clean-edb.RData", wd))
```

```{r q2 sherman mlm setup}
sherman_states_long <- sherman_states_long %>%
  mutate(value2 = value^2)

q2_mlm_sherman_nested <- sherman_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    sherman_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  mutate(p_value_c = p_value - mean(p_value, na.rm = T)
         , s_value_c = s_value - mean(s_value, na.rm = T)) %>%
  group_by(p_name, p_item, s_name) %>%
  nest() %>%
  ungroup()
```

```{r q2 sherman mlm save}
save_fun <- function(d, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/03-sherman/06-q2-mlm/%s-%s-%s.RData", wd, pname, pitem, sname))
}

q2_mlm_sherman_nested %>%
  mutate(d = pmap(list(data, p_name, p_item, s_name), save_fun))
```

```{r}
length(unique((q2_mlm_sherman_nested %>% unnest(data))$SID))
q2_mlm_sherman_nested %>% unnest(data) %>%
  select(SID, Date) %>% distinct() %>%
  group_by(SID) %>%
  tally() %>%
  ungroup() %>% 
  mutate(n = n+1) %>% 
  summarize_at(vars(n), lst(min, max, mean, sd))
```

#### Study 4: Horstmann et al., 2021  

```{r q2 load horstmann}
load(sprintf("%s/03-data/04-horstmann/02-clean-data/horstmann-clean-edb.RData", wd))
```

```{r q2 horstmann mlm setup}
horstmann_states_long <- horstmann_states_long %>%
  mutate(value2 = value^2)

q2_mlm_horstmann_nested <- horstmann_states_long %>% 
  filter(category == "pers") %>%
  select(SID, Date, tdif, p_name = name, p_item = item, p_value = value, p_value2 = value2, p_delta_value = delta_value) %>%
  full_join(
    horstmann_states_long %>% 
      filter(category == "sit") %>%
      select(SID, Date, tdif, s_name = name, s_value = value, s_value2 = value2, s_delta_value = delta_value)
  ) %>%
  filter(!is.na(p_item) & ! is.na(s_name) & !is.na(p_delta_value) & !is.na(s_delta_value)) %>%
  group_by(SID, p_name, p_item, s_name) %>%
  filter(n() >= 19 & sd(p_delta_value) != 0 & sd(s_delta_value) != 0) %>%
  mutate(p_value_c = p_value - mean(p_value, na.rm = T)
         , s_value_c = s_value - mean(s_value, na.rm = T)) %>%
  group_by(p_name, p_item, s_name) %>%
  nest() %>%
  ungroup()
```

```{r q2 horstmann mlm save}
save_fun <- function(d, pname, pitem, sname){
  save(d, file = sprintf("%s/03-data/04-horstmann/06-q2-mlm/%s-%s-%s.RData", wd, pname, pitem, sname))
}

q2_mlm_horstmann_nested %>%
  mutate(d = pmap(list(data, p_name, p_item, s_name), save_fun))
```


```{r}
length(unique((q2_mlm_horstmann_nested %>% unnest(data))$SID))
q2_mlm_horstmann_nested %>% unnest(data) %>%
  select(SID, Date) %>% distinct() %>%
  group_by(SID) %>%
  tally() %>%
  ungroup() %>% 
  mutate(n = n+1) %>% 
  summarize_at(vars(n), lst(min, max, mean, sd))
```

## Models  
### Idiographic  
#### Model Function  
```{r q2 idio model function}
q2_idio_model_fun <- function(study, sid, pname, pitem, sname){
  study2 <- mapvalues(
    study
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , c("pairs", "ipcs", "sherman", "horstmann")
    , warn_missing = F
    )
  load(sprintf("%s/03-data/%s/05-q2-idio/%s-%s-%s-%s.RData", wd, study, sid, pname, pitem, sname))
  
  
  # load sample model 
  load(sprintf("%s/05-results/02-q2/01-idio/sample-idio-linear.RData", wd))
  
  # formula (not added to update because update doesn't support new multivariate formulas)
  # f1 <- bf(p_delta_value ~ 1 + p_value + s_value + p_value:s_value)
  # f2 <- bf(s_delta_value ~ 1 + s_value + p_value + s_value:p_value)
  # f <- mvbrmsformula(f1, f1) + set_rescor(TRUE)
  
  # updated from sample run to prevent compilation and speed runtime
  m <- update(m1
             , newdata = d
             , iter = 2000
             , warmup = 1000
             , cores = 4
             ) 
  save(m, file = sprintf("%s/05-results/02-q2/01-idio/01-models/%s-%s-%s-%s-%s.RData", wd, study2, sid, pname, pitem, sname))
  
  fx <- broom.mixed::tidy(m)
  save(fx, file = sprintf("%s/05-results/02-q2/01-idio/02-summary/%s-%s-%s-%s-%s.RData", wd, study2, sid, pname, pitem, sname))
  
  draws <- as_draws_df(m)
  save(draws, file = sprintf("%s/05-results/02-q2/01-idio/05-draws/%s-%s-%s-%s-%s.RData", wd, study2, sid, pname, pitem, sname))
  
  # find attractor location by solving for x
  # linear               # quadratic
  # y = b0 + b1*x        x = (-b +/ sqrt(b2 - 4ac))/2a
  # 0 = b0 + b1*x        
  # -b0 = b1*x         
  # x = -b0/b1
  equil <- draws %>%
    mutate(
      p__attr_loc_0 = -1 * b_pdeltavalue_Intercept / b_pdeltavalue_p_value
      , p_attr_loc5 = -1*(b_pdeltavalue_Intercept + 5*b_pdeltavalue_s_value)/(b_pdeltavalue_p_value + 5*`b_pdeltavalue_p_value:s_value`)
      , p_attr_loc10 = -1*(b_pdeltavalue_Intercept + 10*b_pdeltavalue_s_value)/(b_pdeltavalue_p_value + 10*`b_pdeltavalue_p_value:s_value`)
      , s__attr_loc_0 = -1 * b_sdeltavalue_Intercept / b_sdeltavalue_s_value
      , s_attr_loc5 = -1*(b_sdeltavalue_Intercept + 5*b_sdeltavalue_p_value)/(b_sdeltavalue_s_value + 5*`b_sdeltavalue_p_value:s_value`)
      , s_attr_loc10 = -1*(b_sdeltavalue_Intercept + 10*b_sdeltavalue_p_value)/(b_sdeltavalue_s_value + 10*`b_sdeltavalue_p_value:s_value`)
      )
  
  # find attractor strength
  equil <- equil %>%
    mutate(
      p__attr_str0    = b_pdeltavalue_p_value
      , p__attr_str5  = b_pdeltavalue_p_value + 5 *`b_pdeltavalue_p_value:s_value`
      , p__attr_str10 = b_pdeltavalue_p_value + 10*`b_pdeltavalue_p_value:s_value`
      , s__attr_str0  = b_sdeltavalue_s_value
      , s__attr_str5  = b_sdeltavalue_s_value + 5 *`b_sdeltavalue_p_value:s_value`
      , s__attr_str10 = b_sdeltavalue_s_value + 10*`b_sdeltavalue_p_value:s_value`
      ) #%>%
    # select(n = .draw, contains("attr_"))
  
  equil <- equil %>% 
    select(contains("attr")) %>%
    posterior_summary() %>%
    as.data.frame() %>% 
    rownames_to_column("term")
  save(equil, file = sprintf("%s/05-results/02-q2/01-idio/03-equil/%s-%s-%s-%s-%s.RData", wd, study2, sid, pname, pitem, sname))
  
  frame <- crossing(p_value = seq(0,10,.5), s_value = seq(0,10,.5))
  
  pred <- bind_cols(
    bind_rows(frame, frame)
    , fitted(m, newdata = frame) %>% array_tree(3) %>% ldply(., .id = "outcome")
    ) %>%
    pivot_wider(
      names_from = "outcome"
      , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
    )
  save(pred, file = sprintf("%s/05-results/02-q2/01-idio/04-predictions/%s-%s-%s-%s-%s.RData", wd, study2, sid, pname, pitem, sname))
  
  # pred %>% 
  #   mutate(col = sqrt(Estimate_pdeltavalue^2 + Estimate_sdeltavalue^2)) %>%
  #   ggplot(aes(x = p_value, y = s_value, u = Estimate_pdeltavalue, v = Estimate_sdeltavalue)) +
  #     geom_quiver(aes(color = col))   +
  #     # geom_hline(aes(yintercept = m_h)) +
  #     # geom_vline(aes(xintercept = m_p)) +
  #     scale_color_gradient(low = "gray80", high = "black") +
  #     labs(x = "Personality State Level", y = "Situation Characteristic Level", title = "") +
  #     theme_classic() +
  #     theme(legend.position = "none",
  #           strip.background = element_rect(fill = "black"),
  #           strip.text = element_text(face = "bold", color = "white", size = rel(1.2)),
  #           axis.text = element_text(face = "bold", color = "black"),
  #           axis.title = element_text(face = "bold", size = rel(1.1)),
  #           plot.title = element_text(face = "bold", hjust = .5))

  rm(list = c("draws", "equil", "m", "m1", "d", "frame", "pred", "fx", "rhs", "f"))
}
```

#### Sample Models  
```{r}
load("/Volumes/Emorie/other projects/ps-oscillators/03-data/02-ipcs/05-q2-idio/01-A-A-Adversity.RData")
# load("03-data/02-ipcs/05-q2-idio/01-A-A-Adversity.RData")

f1 <- bf(p_delta_value ~ 1 + p_value + s_value + p_value:s_value)
f2 <- bf(s_delta_value ~ 1 + p_value + s_value + p_value:s_value)
  
m1 <- brm(mvbrmsformula(f1, f2) + set_rescor(TRUE)
             , data = d
             , iter = 100
             , warmup = 10
             , cores = 4
             )
save(m1, file = sprintf("%s/05-results/02-q2/01-idio/sample-idio-linear.RData", wd))
# save(m1, file = "05-results/02-q2/01-idio/sample-idio-linear.RData")
```


#### Run Models  
```{r q2 idio run models}
done <- tibble(file = list.files(sprintf("%s/05-results/02-q2/01-idio/01-models", wd))) %>%
  separate(file, c("study", "SID", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(study =  mapvalues(study
    , c("pairs", "ipcs", "sherman", "horstmann")
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , warn_missing = F
    )
    , s_name = str_remove_all(s_name, ".RData")
    , done = "done")

q2_idio_nested <- tibble(study = c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")) %>%
  mutate(file = map(study, ~sprintf("%s/03-data/%s/05-q2-idio", wd, .) %>% list.files())) %>%
  unnest(file) %>%
  separate(file, c("SID", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")) %>%
  full_join(done) %>% filter(is.na(done))

plan(multisession(workers = 12L))
q2_idio_nested %>%
  mutate(
    m = future_pmap(
    # m = pmap(
      list(study, SID, category, name, item)
      , .f = possibly(q2_idio_model_fun, NA_real_)
      # , .f = q2_idio_model_fun
      , .progress = T
      , .options = furrr_options(
        globals = c("wd")
        , packages = c("plyr", "tidyverse", "brms")
      )
      )
    )
closeAllConnections()
```

#### Save Arguments For Cluster  
```{r q2 idio save arguments}
q2_idio_nested %>%
  write_delim(., file = sprintf("%s/02-scripts/03-cluster/args/q2-idio.txt", wd))
```

### Multilevel Models  
#### Centered  
##### Model Function  
```{r q2 mlm model function}
q2_mlm_model_fun <- function(study, pname, pitem, sitem){
  study2 <- mapvalues(
    study
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , c("pairs", "ipcs", "sherman", "horstmann")
    , warn_missing = F
    )
  
  # load the data 
  load(sprintf("%s/03-data/%s/06-q2-mlm/%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # setup the formula 
  f1 <- bf(p_delta_value ~ 1 + p_value + s_value + p_value:s_value + (-1 + p_value + s_value + p_value:s_value | p | SID))
  f2 <- bf(s_delta_value ~ 1 + p_value + s_value + p_value:s_value + (-1 + s_value + p_value + p_value:s_value | p | SID))
  
  # load the sample model 
  load(sprintf("%s/05-results/02-q2/02-mlm/sample-mlm-linear.RData", wd))
  
  # run the model using update for faster compilation 
  m <- update(
    m1
    , newdata = d
    , iter = 2000
    , warmup = 1000
    , cores = 4
  )
  save(m, file = sprintf("%s/05-results/02-q2/02-mlm/01-models/%s-%s-%s-%sRData", wd, study2, pname, pitem, sitem))
  
  fx <- broom.mixed::tidy(m)
  rx <- coef(m, probs = c(0.025, 0.975))[[1]] %>% array_tree(3) %>% 
      tibble(term = names(.), data = .) %>% 
      mutate(data = map(data, ~(.) %>% data.frame %>% 
        rownames_to_column("SID"))) %>% 
      unnest(data) %>% 
      select(term, SID, estimate = Estimate, conf.low = Q2.5, conf.high = Q97.5)
  save(fx, rx, file = sprintf("%s/05-results/02-q2/01-idio/02-summary/%s-%s-%s-%sRData", wd, study2, pname, pitem, sitem))
  
  draws <- as_draws_df(m) %>% as_tibble()
  fx_draws <- draws %>% select(-contains("r_SID"), -contains("sd_"), -contains("cor_"))
  rx_draws <- draws %>% 
    mutate(n = 1:n()) %>%
    select(starts_with("r_SID"), n) %>% 
    pivot_longer(
      names_to = "term"
      , values_to = "estimate"
      , cols = -n
    ) %>%
    mutate(term = str_remove_all(term, "r_SID__")) %>%
    separate("term", c("outcome", "term"), sep = "\\[") %>%
    separate("term", c("SID", "term"), sep = ",") %>%
    mutate(term = str_remove_all(term, "\\]")) %>%
    pivot_wider(
      names_from = c("outcome", "term")
      , names_sep = "__"
      , values_from = "estimate"
      ) %>%
    full_join(fx_draws %>% mutate(n = 1:n()) %>% select(n, contains("b_"))) %>%
      mutate(pdeltavalue__p_value_c = pdeltavalue__p_value_c + b_pdeltavalue_p_value_c
             , pdeltavalue__s_value_c = pdeltavalue__s_value_c + b_pdeltavalue_s_value_c
             , `pdeltavalue__p_value_c:s_value_c` = `pdeltavalue__p_value_c:s_value_c` + `b_pdeltavalue_p_value_c:s_value_c`
             , sdeltavalue__p_value_c = sdeltavalue__p_value_c + b_sdeltavalue_p_value_c
             , sdeltavalue__s_value_c = sdeltavalue__s_value_c + b_sdeltavalue_s_value_c
             , `sdeltavalue__p_value_c:s_value_c` = `sdeltavalue__p_value_c:s_value_c` + `b_sdeltavalue_p_value_c:s_value_c`) %>%
      select(-starts_with("b_"))
  
  save(fx_draws, rx_draws
       , file = sprintf("%s/05-results/02-q2/02-mlm/05-draws/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  gr_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>% 
    ungroup() %>% 
    summarize_at(vars(p_value_c, s_value_c), mean) 
  fx_equil <- fx_draws %>% 
    as.data.frame() %>%
    # fixed estimates
    mutate(p__attr_loc_0 = -1*b_pdeltavalue_Intercept/b_pdeltavalue_p_value_c
           , p__attr_loc_p = -1*(b_pdeltavalue_Intercept + gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + gr_sd$s_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(b_pdeltavalue_Intercept + -1*gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c* `b_pdeltavalue_p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*b_sdeltavalue_Intercept/b_sdeltavalue_s_value_c
           , s__attr_loc_p = -1*(b_sdeltavalue_Intercept + gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + gr_sd$p_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(b_sdeltavalue_Intercept + -1*gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c * `b_sdeltavalue_p_value_c:s_value_c`))
  
  # participant-level estimates
  ind_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>%
    ungroup() %>%
    mutate(SID = as.character(SID))
  
  rx_equil <- rx_draws %>%
    full_join(ind_sd) %>%
    # delta personality at 0, 5, 10
    mutate(p__attr_loc_0 = -1*0/pdeltavalue__p_value_c
           , p__attr_loc_p = -1*(0 + s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(0 + -1*s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + -1*s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*0/sdeltavalue__s_value_c
           , s__attr_loc_p = -1*(0 + p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + p_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(0 + -1*p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + -1*p_value_c * `sdeltavalue__p_value_c:s_value_c`))
  
  # find attractor strength
  equil <- fx_equil %>% 
    # fixed estimates
    mutate(
      n = 1:n()
      , SID = "overall"
      , p__attr_str_0 = b_pdeltavalue_p_value_c
      , p__attr_str_p = b_pdeltavalue_p_value_c + gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , p__attr_str_m = b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , s__attr_str_0 = b_sdeltavalue_s_value_c
      , s__attr_str_p = b_sdeltavalue_s_value_c + gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      , s__attr_str_m = b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      ) %>% 
    select(n, SID, contains("attr_loc"), contains("attr_str")) %>%
    full_join(
      # participant-level estimates
      rx_equil %>% 
        mutate(
          SID = as.character(SID)
         , p__attr_str0 = pdeltavalue__p_value_c
         , p__attr_str_p = pdeltavalue__p_value_c + s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , p__attr_str_m = pdeltavalue__p_value_c + -1*s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , s__attr_str0 = sdeltavalue__s_value_c
         , s__attr_str_p = sdeltavalue__s_value_c + p_value_c*`sdeltavalue__p_value_c:s_value_c`
         , s__attr_str_m = sdeltavalue__s_value_c + -1*p_value_c*`sdeltavalue__p_value_c:s_value_c`
        ) #%>%
        # select(n, SID, contains("attr_loc"), contains("attr_str"))
    )
  
  # summarize the posterior draws
  equil <- equil %>% 
    select(SID, contains("attr_loc"), contains("attr_str")) %>%
    group_by(SID) %>%
    nest() %>% 
    ungroup() %>%
    mutate(
      data = map(data
                 , ~(.) %>% 
                   posterior_summary %>%
                   as.data.frame() %>% 
                   rownames_to_column("term")
                )
      ) %>%
    unnest(data)
  save(equil, file = sprintf("%s/05-results/02-q2/02-mlm/03-equil/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # setup predictions for plotting
  # fixed estimates
  gr_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>% 
    ungroup() %>% 
    summarize_at(vars(p_value_c, s_value_c), mean) 
  
  frame <- crossing(p_value_c = seq(-2*gr_sd$p_value_c, 2*gr_sd$p_value_c, length.out = 25), s_value_c = seq(-2*gr_sd$s_value_c, 2*gr_sd$s_value_c, length.out = 25))
  fx_pred <- bind_cols(
    bind_rows(frame, frame)
    , fitted(m, newdata = frame, re_formula = NA) %>% array_tree(3) %>% ldply(., .id = "outcome")
  ) %>%
    pivot_wider(
      names_from = "outcome"
      , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
    )
  
  crossing_fun <- function(d){
    crossing(
      p_value_c = seq(-2*d$p_value_c, 2*d$p_value_c, length.out = 25)
      , s_value_c = seq(-2*d$s_value_c, 2*d$s_value_c, length.out = 25)
      )
  }
  # participant-level estimates
  frame <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>%
    ungroup() %>%
    mutate(SID = as.character(SID)) %>%
    group_by(SID) %>%
    nest() %>%
    ungroup() %>%
    mutate(data = map(data, crossing_fun)) %>%
    unnest(data)
  
  # frame <- crossing(SID = unique(m$data$SID), p_value_c = seq(-5,5,.25), s_value_c = seq(-5,5,.25))
  rx_pred <- bind_cols(
    bind_rows(frame, frame)
    , fitted(m, newdata = frame) %>% array_tree(3) %>% ldply(., .id = "outcome")
  ) %>%
    pivot_wider(
      names_from = "outcome"
      , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
    )
  save(fx_pred, rx_pred
       , file = sprintf("05-results/02-q2/02-mlm/04-predictions/%s-%s-%s-%s.RData", study, pname, pitem, sitem))
  
  # clean up and return nothing
  rm(list = c("frame", "fx_pred", "m1", "m", "fx", "rx", "fx_draws", "rx_draws", "equil", "draws"))
  gc()
  return(T)
}
```

##### Sample Models  
```{r q2 mlm sample model}
load("/Volumes/Emorie/other projects/ps-oscillators/03-data/02-ipcs/06-q2-mlm/A-A-Adversity.RData")
# load("03-data/02-ipcs/06-q2-mlm/A-A-Adversity.RData")

f1 <- bf(p_delta_value ~ 1 + p_value_c + s_value_c + p_value_c:s_value_c + (-1 + p_value_c + s_value_c + p_value_c:s_value_c | p | SID))
f2 <- bf(s_delta_value ~ 1 + p_value_c + s_value_c + p_value_c:s_value_c + (-1 + p_value_c + s_value_c + p_value_c:s_value_c | p | SID))

m1 <- brm(
    mvbrmsformula(f1, f2) + set_rescor(TRUE)
    , data = d
    , iter = 100
    , warmup = 10
    # , cores = 4
    , cores = 12
    , threads = threading(3)
    , backend = "cmdstanr"
    , chains = 4
  )
save(m1, file = sprintf("%s/05-results/02-q2/02-mlm/sample-mlm-linear.RData", wd))
# save(m1, file = "05-results/02-q2/02-mlm/sample-mlm-linear.RData")
```


##### Run Models  
```{r q2 mlm run models}
done <- tibble(file = list.files(sprintf("%s/05-results/02-q2/02-mlm/01-models", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(study =  mapvalues(study
    , c("pairs", "ipcs", "sherman", "horstmann")
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , warn_missing = F
    )
    , s_name = str_remove_all(s_name, ".RData")
    , done = "done")

q2_mlm_nested <- tibble(study = c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")) %>%
  mutate(file = map(study, ~sprintf("%s/03-data/%s/06-q2-mlm", wd, .) %>% list.files())) %>%
  unnest(file) %>%
  separate(file, c("p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")) %>%
  full_join(done) %>% filter(is.na(done))

plan(multisession(workers = 12L))
q2_mlm_nested %>%
  mutate(
    m = future_pmap(
    # m = pmap(
      list(study, p_name, p_item, s_name)
      # , .f = possibly(q2_mlm_model_fun, NA_real_)
      , .f = q2_mlm_model_fun
      , .progress = T
      , .options = furrr_options(
        globals = c("wd")
        , packages = c("plyr", "tidyverse", "brms")
      )
      )
    )
closeAllConnections()
```

##### Save Arguments For Cluster  
```{r q2 mlm save arguments}
q2_mlm_nested %>%
  write_delim(., file = sprintf("%s/02-scripts/03-cluster/args/q2-mlm.txt", wd))
```

##### DELETE LATER  
```{r}
q2_pred_fun <- function(study, pname, pitem, sitem){
  load(sprintf("%s/05-results/02-q2/02-mlm/01-models/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  load(sprintf("%s/05-results/02-q2/02-mlm/05-draws/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # find attractor location by solving for x
  # linear               # quadratic
  # y = b0 + b1*x        x = (-b +/ sqrt(b2 - 4ac))/2a
  # 0 = b0 + b1*x        
  # -b0 = b1*x         
  # x = -b0/b1
  gr_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>% 
    ungroup() %>% 
    summarize_at(vars(p_value_c, s_value_c), mean) 
  fx_equil <- fx_draws %>% 
    as.data.frame() %>%
    # fixed estimates
    mutate(p__attr_loc_0 = -1*b_pdeltavalue_Intercept/b_pdeltavalue_p_value_c
           , p__attr_loc_p = -1*(b_pdeltavalue_Intercept + gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + gr_sd$s_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(b_pdeltavalue_Intercept + -1*gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c* `b_pdeltavalue_p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*b_sdeltavalue_Intercept/b_sdeltavalue_s_value_c
           , s__attr_loc_p = -1*(b_sdeltavalue_Intercept + gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + gr_sd$p_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(b_sdeltavalue_Intercept + -1*gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c * `b_sdeltavalue_p_value_c:s_value_c`))
  
  # participant-level estimates
  ind_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>%
    ungroup() %>%
    mutate(SID = as.character(SID))
  
  rx_equil <- rx_draws %>%
    full_join(ind_sd) %>%
    # delta personality at 0, 5, 10
    mutate(p__attr_loc_0 = -1*0/pdeltavalue__p_value_c
           , p__attr_loc_p = -1*(0 + s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(0 + -1*s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + -1*s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*0/sdeltavalue__s_value_c
           , s__attr_loc_p = -1*(0 + p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + p_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(0 + -1*p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + -1*p_value_c * `sdeltavalue__p_value_c:s_value_c`))
  
  # find attractor strength
  equil <- fx_equil %>% 
    # fixed estimates
    mutate(
      n = 1:n()
      , SID = "overall"
      , p__attr_str_0 = b_pdeltavalue_p_value_c
      , p__attr_str_p = b_pdeltavalue_p_value_c + gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , p__attr_str_m = b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , s__attr_str_0 = b_sdeltavalue_s_value_c
      , s__attr_str_p = b_sdeltavalue_s_value_c + gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      , s__attr_str_m = b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      ) %>% 
    select(n, SID, contains("attr_loc"), contains("attr_str")) %>%
    full_join(
      # participant-level estimates
      rx_equil %>% 
        mutate(
          SID = as.character(SID)
         , p__attr_str0 = pdeltavalue__p_value_c
         , p__attr_str_p = pdeltavalue__p_value_c + s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , p__attr_str_m = pdeltavalue__p_value_c + -1*s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , s__attr_str0 = sdeltavalue__s_value_c
         , s__attr_str_p = sdeltavalue__s_value_c + p_value_c*`sdeltavalue__p_value_c:s_value_c`
         , s__attr_str_m = sdeltavalue__s_value_c + -1*p_value_c*`sdeltavalue__p_value_c:s_value_c`
        ) #%>%
        # select(n, SID, contains("attr_loc"), contains("attr_str"))
    )
  
  # summarize the posterior draws
  equil <- equil %>% 
    select(SID, contains("attr_loc"), contains("attr_str")) %>%
    group_by(SID) %>%
    nest() %>% 
    ungroup() %>%
    mutate(
      data = map(data
                 , ~(.) %>% 
                   posterior_summary %>%
                   as.data.frame() %>% 
                   rownames_to_column("term")
                )
      ) %>%
    unnest(data)
  save(equil, file = sprintf("%s/05-results/02-q2/02-mlm/03-equil/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  
  # participant-level estimates
  # frame <- crossing(SID = unique(m$data$SID), p_value_c = seq(0,10,.5), s_value_c = seq(0,10,.5))
  # rx_pred <- bind_cols(
  #   bind_rows(frame, frame)
  #   , fitted(m, newdata = frame) %>% array_tree(3) %>% ldply(., .id = "outcome")
  #   ) %>%
  #   pivot_wider(
  #     names_from = "outcome"
  #     , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
  #   )
  # save(fx_pred, rx_pred
  #      , file = sprintf("%s/05-results/02-q2/01-idio/04-predictions/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # clean up and return nothing
  rm(list = c("frame", "fx_pred", "m1", "m", "fx", "rx", "fx_draws", "rx_draws", "equil", "draws"))
  gc()
  return(T)
}

done <- tibble(file = list.files(sprintf("%s/05-results/02-q2/02-mlm/04-predictions", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")
    , done = "done")

q2_mlm_nested <- tibble(file = list.files(sprintf("%s/05-results/02-q2/02-mlm/01-models", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")) %>%
  full_join(done) %>% filter(is.na(done))

plan(multisession(workers = 8L))
q2_mlm_nested %>%
  filter(study == "sherman") %>%
  mutate(
    m = future_pmap(
    # m = pmap(
      list(study, p_name, p_item, s_name)
      , .f = possibly(q2_pred_fun, NA_real_)
      # , .f = q2_pred_fun
      , .progress = T
      , .options = furrr_options(
        globals = c("wd")
        , packages = c("plyr", "tidyverse", "brms")
      )
      )
    )
closeAllConnections()

q2_mlm_nested %>%
  write_delim(., file = sprintf("%s/02-scripts/03-cluster/args/q2-mlm-pred.txt", wd))
```

#### Uncentered  
##### Model Function  
```{r q2 mlm model function}
q2_mlm_model_fun <- function(study, pname, pitem, sitem){
  study2 <- mapvalues(
    study
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , c("pairs", "ipcs", "sherman", "horstmann")
    , warn_missing = F
    )
  
  # load the data 
  load(sprintf("%s/03-data/%s/06-q2-mlm/%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # setup the formula 
  f1 <- bf(p_delta_value ~ 1 + p_value + s_value + p_value:s_value + (1 + p_value + s_value + p_value:s_value | p | SID))
  f2 <- bf(s_delta_value ~ 1 + p_value + s_value + p_value:s_value + (1 + s_value + p_value + p_value:s_value | p | SID))
  
  # load the sample model 
  load(sprintf("%s/05-results/02-q2/03-mlm-uncentered/sample-mlm-linear.RData", wd))
  
  Prior <- c(prior(cauchy(0, .1), class = sd, resp = "pdeltavalue"),
           prior(cauchy(0, .1), class = sd, resp = "sdeltavalue"),
           prior(exponential(1), class = sigma, resp = "pdeltavalue"),
           prior(exponential(1), class = sigma, resp = "sdeltavalue"),
           prior(lkj(1), class = cor))
  
  # run the model using update for faster compilation 
  m <- update(
    m1
    , newdata = d
    , iter = 200
    , warmup = 100
    , cores = 4
    # , prior = Prior
  )
  save(m, file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/01-models/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  fx <- broom.mixed::tidy(m)
  rx <- coef(m, probs = c(0.025, 0.975))[[1]] %>% array_tree(3) %>% 
      tibble(term = names(.), data = .) %>% 
      mutate(data = map(data, ~(.) %>% data.frame %>% 
        rownames_to_column("SID"))) %>% 
      unnest(data) %>% 
      select(term, SID, estimate = Estimate, conf.low = Q2.5, conf.high = Q97.5)
  save(fx, rx, file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/02-summary/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  draws <- as_draws_df(m) %>% as_tibble()
  fx_draws <- draws %>% select(-contains("r_SID"), -contains("sd_"), -contains("cor_"))
  rx_draws <- draws %>% 
    mutate(n = 1:n()) %>%
    select(starts_with("r_SID"), n) %>% 
    pivot_longer(
      names_to = "term"
      , values_to = "estimate"
      , cols = -n
    ) %>%
    mutate(term = str_remove_all(term, "r_SID__")) %>%
    separate("term", c("outcome", "term"), sep = "\\[") %>%
    separate("term", c("SID", "term"), sep = ",") %>%
    mutate(term = str_remove_all(term, "\\]")) %>%
    mutate(term = str_replace_all(term, "s_value:p_value", "p_value:s_value")) %>%
    pivot_wider(
      names_from = c("outcome", "term")
      , names_sep = "__"
      , values_from = "estimate"
      ) %>%
    full_join(fx_draws %>% mutate(n = 1:n()) %>% select(n, contains("b_"))) %>%
      mutate(pdeltavalue__Intercept = pdeltavalue__Intercept + b_pdeltavalue_Intercept
             , pdeltavalue__p_value = pdeltavalue__p_value + b_pdeltavalue_p_value
             , pdeltavalue__s_value = pdeltavalue__s_value + b_pdeltavalue_s_value
             , `pdeltavalue__p_value:s_value` = `pdeltavalue__p_value:s_value` + `b_pdeltavalue_p_value:s_value`
             , sdeltavalue__Intercept = sdeltavalue__Intercept + b_sdeltavalue_Intercept
             , sdeltavalue__p_value = sdeltavalue__p_value + b_sdeltavalue_p_value
             , sdeltavalue__s_value = sdeltavalue__s_value + b_sdeltavalue_s_value
             , `sdeltavalue__p_value:s_value` = `sdeltavalue__p_value:s_value` + `b_sdeltavalue_p_value:s_value`) %>%
      select(-starts_with("b_"))
  save(fx_draws, rx_draws
       , file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/05-draws/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  # find attractor location by solving for x
  # linear               # quadratic
  # y = b0 + b1*x        x = (-b +/ sqrt(b2 - 4ac))/2a
  # 0 = b0 + b1*x        
  # -b0 = b1*x         
  # x = -b0/b1
  fx_equil <- fx_draws %>% 
    as.data.frame() %>%
    # fixed estimates
    mutate(p__attr_loc_0 = -1*b_pdeltavalue_Intercept/b_pdeltavalue_p_value
           , p__attr_loc5 = -1*(b_pdeltavalue_Intercept + 5*b_pdeltavalue_s_value)/(b_pdeltavalue_p_value + 5 * `b_pdeltavalue_p_value:s_value`)
           , p__attr_loc10 = -1*(b_pdeltavalue_Intercept + 10*b_pdeltavalue_s_value)/(b_pdeltavalue_p_value + 10 * `b_pdeltavalue_p_value:s_value`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*b_sdeltavalue_Intercept/b_sdeltavalue_s_value
           , s__attr_loc5 = -1*(b_sdeltavalue_Intercept + 5*b_sdeltavalue_p_value)/(b_sdeltavalue_s_value + 5 * `b_pdeltavalue_p_value:s_value`)
           , s__attr_loc10 = -1*(b_sdeltavalue_Intercept + 10*b_sdeltavalue_p_value)/(b_sdeltavalue_s_value + 10 * `b_sdeltavalue_p_value:s_value`))
  # participant-level estimates
  rx_equil <- rx_draws %>%
    # delta personality at 0, 5, 10
    mutate(p__attr_loc_0 = -1*pdeltavalue__Intercept/pdeltavalue__p_value
           , p__attr_loc5 = -1*(pdeltavalue__Intercept + 5*pdeltavalue__s_value)/(pdeltavalue__p_value + 5 * `pdeltavalue__p_value:s_value`)
           , p__attr_loc10 = -1*(pdeltavalue__Intercept + 10*pdeltavalue__s_value)/(pdeltavalue__p_value + 10 * `pdeltavalue__p_value:s_value`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*sdeltavalue__Intercept/sdeltavalue__s_value
           , s__attr_loc5 = -1*(sdeltavalue__Intercept + 5*sdeltavalue__p_value)/(sdeltavalue__s_value + 5 * `pdeltavalue__p_value:s_value`)
           , s__attr_loc10 = -1*(sdeltavalue__Intercept + 10*sdeltavalue__p_value)/(sdeltavalue__s_value + 10 * `sdeltavalue__p_value:s_value`))
  
  # find attractor strength
  equil <- fx_equil %>% 
    # fixed estimates
    mutate(
      n = 1:n()
      , SID = "overall"
      , p__attr_str0 = b_pdeltavalue_p_value
      , p__attr_str5 = b_pdeltavalue_p_value + 5*`b_pdeltavalue_p_value:s_value`
      , p__attr_str10 = b_pdeltavalue_p_value + 10*`b_pdeltavalue_p_value:s_value`
      , s__attr_str0 = b_sdeltavalue_s_value
      , s__attr_str5 = b_sdeltavalue_s_value + 5*`b_sdeltavalue_p_value:s_value`
      , s__attr_str10 = b_sdeltavalue_s_value + 10*`b_sdeltavalue_p_value:s_value`
      ) %>% 
    select(n, SID, contains("attr_loc"), contains("attr_str")) %>%
    full_join(
      # participant-level estimates
      rx_equil %>% 
        mutate(
         p__attr_str0 = pdeltavalue__p_value
         , p__attr_str5 = pdeltavalue__p_value + 5*`pdeltavalue__p_value:s_value`
         , p__attr_str10 = pdeltavalue__p_value + 10*`pdeltavalue__p_value:s_value`
         , s__attr_str0 = sdeltavalue__s_value
         , s__attr_str5 = sdeltavalue__s_value + 5*`sdeltavalue__p_value:s_value`
         , s__attr_str10 = sdeltavalue__s_value + 10*`sdeltavalue__p_value:s_value`
        ) #%>%
        # select(n, SID, contains("attr_loc"), contains("attr_str"))
    )
  
  # summarize the posterior draws
  equil <- equil %>% 
    select(SID, contains("attr_loc"), contains("attr_str")) %>%
    group_by(SID) %>%
    nest() %>% 
    ungroup() %>%
    mutate(
      data = map(data
                 , ~(.) %>% 
                   posterior_summary %>%
                   as.data.frame() %>% 
                   rownames_to_column("term")
                )
      ) %>%
    unnest(data)
  save(equil, file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/03-equil/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  # setup predictions for plotting
  # fixed estimates
  frame <- crossing(p_value = seq(0,10,.5), s_value = seq(0,10,.5))
  fx_pred <- bind_cols(
    bind_rows(frame, frame)
    , fitted(m, newdata = frame, re_formula = NA) %>% array_tree(3) %>% ldply(., .id = "outcome")
    ) %>%
    pivot_wider(
      names_from = "outcome"
      , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
    )
  # participant-level estimates
  frame <- crossing(SID = unique(m$data$SID), p_value = seq(0,10,.5), s_value = seq(0,10,.5))
  rx_pred <- bind_cols(
    bind_rows(frame, frame)
    , fitted(m, newdata = frame) %>% array_tree(3) %>% ldply(., .id = "outcome")
    ) %>%
    pivot_wider(
      names_from = "outcome"
      , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
    )
  save(fx_pred, rx_pred
       , file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/04-predictions/%s-%s-%s-%s.RData", wd, study2, pname, pitem, sitem))
  
  # clean up and return nothing
  rm(list = c("frame", "fx_pred", "m1", "m", "fx", "rx", "fx_draws", "rx_draws", "equil", "draws"))
  gc()
  return(T)
}
```

##### Sample Models  
```{r q2 mlm sample model}
##### Question 2: MLM ######

# setwd("/projects/p31385/ps-oscillators")
# 
# library(rstan)
# library(plyr)
# library(tidyverse)
# library(psych)
# library(brms)
load("/Volumes/Emorie/other projects/ps-oscillators/03-data/02-ipcs/06-q2-mlm/A-A-Adversity.RData")
# load("03-data/02-ipcs/06-q2-mlm/A-A-Adversity.RData")

f1 <- bf(p_delta_value ~ 1 + p_value + s_value + p_value:s_value + (1 + p_value + s_value + p_value:s_value | p | SID))
f2 <- bf(s_delta_value ~ 1 + p_value + s_value + p_value:s_value + (1 + p_value + s_value + p_value:s_value | p | SID))

Prior <- c(prior(cauchy(0, .1), class = sd, resp = "pdeltavalue"),
           prior(cauchy(0, .1), class = sd, resp = "sdeltavalue"),
           prior(exponential(1), class = sigma, resp = "pdeltavalue"),
           prior(exponential(1), class = sigma, resp = "sdeltavalue"),
           prior(lkj(1), class = cor))

m1 <- brm(
    mvbrmsformula(f1, f2) + set_rescor(TRUE)
    , data = d
    , iter = 100
    , warmup = 10
    , cores = 4
    , prior = Prior
  )
save(m1, file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/sample-mlm-linear.RData", wd))
# save(m1, file = "05-results/02-q2/03-mlm-uncentered/sample-mlm-linear.RData")
```


##### Run Models  
```{r q2 mlm run models}
done <- tibble(file = list.files(sprintf("%s/05-results/02-q2/03-mlm-uncentered/01-models", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(study =  mapvalues(study
    , c("pairs", "ipcs", "sherman", "horstmann")
    , c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")
    , warn_missing = F
    )
    , s_name = str_remove_all(s_name, ".RData")
    , done = "done")

q2_mlm_nested <- tibble(study = c("01-pairs", "02-ipcs", "03-sherman", "04-horstmann")) %>%
  mutate(file = map(study, ~sprintf("%s/03-data/%s/06-q2-mlm", wd, .) %>% list.files())) %>%
  unnest(file) %>%
  separate(file, c("p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")) %>%
  full_join(done) %>% filter(is.na(done))

plan(multisession(workers = 3L))
q2_mlm_nested %>%
  filter(study == "02-ipcs" & p_name == p_item) %>%
  mutate(
    m = future_pmap(
    # m = pmap(
      list(study, p_name, p_item, s_name)
      # , .f = possibly(q2_mlm_model_fun, NA_real_)
      , .f = q2_mlm_model_fun
      , .progress = T
      , .options = furrr_options(
        globals = c("wd")
        , packages = c("plyr", "tidyverse", "brms")
      )
      )
    )
closeAllConnections()
```

##### Save Arguments For Cluster  
```{r q2 mlm save arguments}
q2_mlm_nested %>%
  filter(p_name == p_item) %>%
  write_delim(., file = sprintf("%s/02-scripts/03-cluster/args/q2-mlm-uncentered.txt", wd))
```

##### DELETE LATER  
```{r}
q2_pred_fun <- function(study, pname, pitem, sitem){
  load(sprintf("%s/05-results/02-q2/02-mlm/01-models/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  load(sprintf("%s/05-results/02-q2/02-mlm/05-draws/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # find attractor location by solving for x
  # linear               # quadratic
  # y = b0 + b1*x        x = (-b +/ sqrt(b2 - 4ac))/2a
  # 0 = b0 + b1*x        
  # -b0 = b1*x         
  # x = -b0/b1
  gr_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>% 
    ungroup() %>% 
    summarize_at(vars(p_value_c, s_value_c), mean) 
  fx_equil <- fx_draws %>% 
    as.data.frame() %>%
    # fixed estimates
    mutate(p__attr_loc_0 = -1*b_pdeltavalue_Intercept/b_pdeltavalue_p_value_c
           , p__attr_loc_p = -1*(b_pdeltavalue_Intercept + gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + gr_sd$s_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(b_pdeltavalue_Intercept + -1*gr_sd$s_value_c*b_pdeltavalue_s_value_c)/(b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c* `b_pdeltavalue_p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*b_sdeltavalue_Intercept/b_sdeltavalue_s_value_c
           , s__attr_loc_p = -1*(b_sdeltavalue_Intercept + gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + gr_sd$p_value_c * `b_pdeltavalue_p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(b_sdeltavalue_Intercept + -1*gr_sd$p_value_c*b_sdeltavalue_p_value_c)/(b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c * `b_sdeltavalue_p_value_c:s_value_c`))
  
  # participant-level estimates
  ind_sd <- m$data %>% 
    group_by(SID) %>% 
    summarize_at(vars(p_value_c, s_value_c), sd) %>%
    ungroup() %>%
    mutate(SID = as.character(SID))
  
  rx_equil <- rx_draws %>%
    full_join(ind_sd) %>%
    # delta personality at 0, 5, 10
    mutate(p__attr_loc_0 = -1*0/pdeltavalue__p_value_c
           , p__attr_loc_p = -1*(0 + s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , p__attr_loc_m = -1*(0 + -1*s_value_c*pdeltavalue__s_value_c)/(pdeltavalue__p_value_c + -1*s_value_c * `pdeltavalue__p_value_c:s_value_c`)
           # delta situations at 0, 5, 10
           , s__attr_loc_0 = -1*0/sdeltavalue__s_value_c
           , s__attr_loc_p = -1*(0 + p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + p_value_c * `pdeltavalue__p_value_c:s_value_c`)
           , s__attr_loc_m = -1*(0 + -1*p_value_c*sdeltavalue__p_value_c)/(sdeltavalue__s_value_c + -1*p_value_c * `sdeltavalue__p_value_c:s_value_c`))
  
  # find attractor strength
  equil <- fx_equil %>% 
    # fixed estimates
    mutate(
      n = 1:n()
      , SID = "overall"
      , p__attr_str_0 = b_pdeltavalue_p_value_c
      , p__attr_str_p = b_pdeltavalue_p_value_c + gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , p__attr_str_m = b_pdeltavalue_p_value_c + -1*gr_sd$s_value_c*`b_pdeltavalue_p_value_c:s_value_c`
      , s__attr_str_0 = b_sdeltavalue_s_value_c
      , s__attr_str_p = b_sdeltavalue_s_value_c + gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      , s__attr_str_m = b_sdeltavalue_s_value_c + -1*gr_sd$p_value_c*`b_sdeltavalue_p_value_c:s_value_c`
      ) %>% 
    select(n, SID, contains("attr_loc"), contains("attr_str")) %>%
    full_join(
      # participant-level estimates
      rx_equil %>% 
        mutate(
          SID = as.character(SID)
         , p__attr_str0 = pdeltavalue__p_value_c
         , p__attr_str_p = pdeltavalue__p_value_c + s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , p__attr_str_m = pdeltavalue__p_value_c + -1*s_value_c*`pdeltavalue__p_value_c:s_value_c`
         , s__attr_str0 = sdeltavalue__s_value_c
         , s__attr_str_p = sdeltavalue__s_value_c + p_value_c*`sdeltavalue__p_value_c:s_value_c`
         , s__attr_str_m = sdeltavalue__s_value_c + -1*p_value_c*`sdeltavalue__p_value_c:s_value_c`
        ) #%>%
        # select(n, SID, contains("attr_loc"), contains("attr_str"))
    )
  
  # summarize the posterior draws
  equil <- equil %>% 
    select(SID, contains("attr_loc"), contains("attr_str")) %>%
    group_by(SID) %>%
    nest() %>% 
    ungroup() %>%
    mutate(
      data = map(data
                 , ~(.) %>% 
                   posterior_summary %>%
                   as.data.frame() %>% 
                   rownames_to_column("term")
                )
      ) %>%
    unnest(data)
  save(equil, file = sprintf("%s/05-results/02-q2/02-mlm/03-equil/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  
  # participant-level estimates
  # frame <- crossing(SID = unique(m$data$SID), p_value_c = seq(0,10,.5), s_value_c = seq(0,10,.5))
  # rx_pred <- bind_cols(
  #   bind_rows(frame, frame)
  #   , fitted(m, newdata = frame) %>% array_tree(3) %>% ldply(., .id = "outcome")
  #   ) %>%
  #   pivot_wider(
  #     names_from = "outcome"
  #     , values_from = c("Estimate", "Est.Error", "Q2.5", "Q97.5")
  #   )
  # save(fx_pred, rx_pred
  #      , file = sprintf("%s/05-results/02-q2/01-idio/04-predictions/%s-%s-%s-%s.RData", wd, study, pname, pitem, sitem))
  
  # clean up and return nothing
  rm(list = c("frame", "fx_pred", "m1", "m", "fx", "rx", "fx_draws", "rx_draws", "equil", "draws"))
  gc()
  return(T)
}

done <- tibble(file = list.files(sprintf("%s/05-results/02-q2/02-mlm/04-predictions", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")
    , done = "done")

q2_mlm_nested <- tibble(file = list.files(sprintf("%s/05-results/02-q2/03-mlm-uncentered/01-models", wd))) %>%
  separate(file, c("study", "p_name", "p_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove_all(s_name, ".RData")) %>%
  full_join(done) %>% filter(is.na(done))

plan(multisession(workers = 8L))
q2_mlm_nested %>%
  filter(study == "sherman") %>%
  mutate(
    m = future_pmap(
    # m = pmap(
      list(study, p_name, p_item, s_name)
      , .f = possibly(q2_pred_fun, NA_real_)
      # , .f = q2_pred_fun
      , .progress = T
      , .options = furrr_options(
        globals = c("wd")
        , packages = c("plyr", "tidyverse", "brms")
      )
      )
    )
closeAllConnections()

q2_mlm_nested %>%
  filter(study == "ipcs") %>%
  write_delim(., file = sprintf("%s/02-scripts/03-cluster/args/q2-mlm-pred.txt", wd))
```


## Compile Results  
### Idiographic  


```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/01-idio/02-summary/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

plan(multisession(workers = 16L))
q2_idio_fx <- tibble(file = sprintf("%s/05-results/02-q2/01-idio/02-summary", wd) %>% list.files()) %>%
  separate(file, c("study", "SID", "t_name", "t_item", "s_name"), sep = "-", remove = F) %>%
  filter(t_name == t_item & !is.na(s_name)) %>%
  mutate(fx = future_map2(
    file, "fx"
    , possibly(loadRData, NA_real_)
    , .progress = T
    , .options = furrr_options(
      globals = c("wd")
      , packages = c("plyr", "tidyverse", "brms")
    )
    )) %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  select(-file)
closeAllConnections()
```

```{r}

```

```{r}
cor_fun <- function(d){
  r <- cor.ci(cbind(d$pdeltavalue, d$sdeltavalue), plot = F)
  tibble(cor = r$rho[2,1], conf.low = r$ci$lower, conf.high = r$ci$upper)
}

q2_idio_dx_cor <- q2_idio_fx %>% 
  unnest(fx) %>%
  filter(term == "(Intercept)") %>% 
  select(study, SID, t_name, s_name, response, estimate) %>% 
  pivot_wider(
    names_from = "response"
    , values_from = "estimate"
  ) %>%
  group_by(study, t_name, s_name) %>% 
  nest() %>%
  ungroup() %>%
  mutate(r = map(data, cor_fun)) %>%
  unnest(r)
```


```{r}
q2_idio_fix_sig <- q2_idio_fx %>% 
  filter(!is.na(fx)) %>%
  unnest(fx) %>% 
  filter(grepl(":", term)) %>%
  mutate(sig = ifelse(sign(conf.low) == sign(conf.high), "sig", "ns")) %>%
  left_join(expected) %>%
  filter(expy == "yes" & sig == "sig") %>%
  select(study, SID, t_name, t_item, s_name) %>%
  distinct()
```

#### Figures  

##### Vector Fields  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/01-idio/04-predictions/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_idio_pred <- tibble(file = sprintf("%s/05-results/02-q2/01-idio/04-predictions", wd) %>% list.files()) %>%
  separate(file, c("study", "SID", "t_name", "t_item", "s_name"), sep = "-", remove = F) %>%
  filter(t_name == t_item & !is.na(s_name)) %>%
  mutate(fx_pred = map2(file, "pred", possibly(loadRData, NA_real_))) %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>% 
  select(-file)
```

```{r}
q2_idio_vec_fun <- function(d, study, sid, sname){
  sname2 <- mapvalues(sname, cat_names$name_old, cat_names$name_new, warn_missing = F)
  d %>% 
    mutate(t_name = factor(t_name, cat_names$name_old, cat_names$name_new)) %>%
    # filter(p_value %in% seq(0,10,1) & s_value %in% seq(0,10,1)) %>%
    mutate(col = sqrt(Estimate_pdeltavalue^2 + Estimate_sdeltavalue^2)) %>%
    ggplot(aes(x = p_value, y = s_value, u = Estimate_pdeltavalue, v = Estimate_sdeltavalue)) + 
    # geom_point(data = df4, color = "white") +
    geom_quiver(rescale = T)   +
    # geom_quiver(aes(color = col))   +
    scale_color_gradient(low = "gray60", high = "black")+
    labs(x = "Personality State Level", y = sprintf("%s Level", sname2)) +
    facet_wrap(~t_name, scales = "free", ncol = 3) +
    theme_classic() +
    theme(legend.position = "none",
          strip.background = element_rect(fill = "black"),
          strip.text = element_text(face = "bold", color = "white", size = rel(1.2)),
          axis.text = element_text(face = "bold", color = "black"),
          axis.title = element_text(face = "bold", size = rel(1.1)))
  ggsave(file = sprintf("%s/05-results/04-plots/02-q2/01-idio/01-vector-fields/%s-%s-%s.png", wd, study, sid, sname)
         , width = 12, height = 8)
  ggsave(file = sprintf("%s/05-results/04-plots/02-q2/01-idio/01-vector-fields/pdf/%s-%s-%s.pdf", wd, study, sid, sname)
         , width = 12, height = 8)
}

q2_idio_pred_p <- q2_idio_pred %>%
  unnest(fx_pred) %>%
  right_join(q2_idio_fix_sig) %>%
  group_by(study, SID, s_name, t_name) %>%
  filter(study == "ipcs") %>%
  nest() %>%
  ungroup() %>%
  mutate(p = pmap(list(data, study, SID, s_name), q2_idio_vec_fun))
```

```{r}
q2_idio_vec_sig_fun <- function(d, study, sid, tname, sname){
  sname2 <- mapvalues(sname, cat_names$name_old, cat_names$name_new, warn_missing = F)
  tname2 <- mapvalues(tname, cat_names$name_old, cat_names$name_new, warn_missing = F)
  study2 <- mapvalues(study, c("ipcs", "sherman", "horstmann"), c("Sample #1\n(Beck & Jackson)", "Sample #2\n(Sherman et al.)", "Sample #3\n(Horstmann et al.)"), warn_missing = F)
  p <- d %>% 
    filter(p_value %in% seq(0,10,1) & s_value %in% seq(0,10,1)) %>%
    mutate(col = sqrt(Estimate_pdeltavalue^2 + Estimate_sdeltavalue^2)) %>%
    # ggplot(aes(x = p_value, y = s_value)) +
    # geom_segment(aes(xend = p_value+Estimate_pdeltavalue/3, yend = s_value+Estimate_sdeltavalue/3)
    #              ,arrow = arrow(length = unit(0.2, "cm")), size = 0.5) +
    ggplot(aes(x = p_value, y = s_value, u = Estimate_pdeltavalue, v = Estimate_sdeltavalue)) +
    geom_quiver(rescale = F) +
    # geom_quiver(aes(color = col)) +
    # scale_color_gradient(low = "gray60", high = "black") +
    labs(x = sprintf("%s Level", tname2), y = sprintf("%s Level", sname2)
         , title = sprintf("%s: %s", study2, sid)) +
    theme_classic() +
    theme(legend.position = "none"
          # , strip.background = element_rect(fill = "black")
          # , strip.text = element_text(face = "bold", color = "white", size = rel(1.2))
          , axis.text = element_text(face = "bold", color = "black")
          , axis.title = element_text(face = "bold", size = rel(1.1))
          , plot.title = element_text(face = "bold", hjust = .5)
        )
  ggsave(p, file = sprintf("%s/05-results/04-plots/02-q2/01-idio/02-sig-vector-fields/%s-%s-%s-%s.png", wd, study, sid, tname, sname)
         , width = 3, height = 3)
  ggsave(p, file = sprintf("%s/05-results/04-plots/02-q2/01-idio/02-sig-vector-fields/pdf/%s-%s-%s-%s.pdf", wd, study, sid, tname, sname)
         , width = 3, height = 3)
  return(p)
}

plan(multisession(workers = 16L))
q2_idio_pred_p <- q2_idio_pred %>%
  unnest(fx_pred) %>%
  right_join(q2_idio_fix_sig) %>%
  group_by(study, SID, s_name, t_name) %>%
  # filter(study == "ipcs") %>%
  nest() %>%
  ungroup() %>%
  mutate(p = future_pmap(
    list(data, study, SID, s_name, t_name)
    , q2_idio_vec_sig_fun
    , .progress = T
    , .options = furrr_options(
      globals = c("wd", "cat_names")
      , packages = c("plyr", "tidyverse", "brms", "ggridges", "ggquiver")
    )
    ))
closeAllConnections()
```

```{r q2 idio gifs}
setwd(sprintf("%s/05-results/04-plots/02-q2/01-idio/03-vector-field-gifs", wd))
library(animation)
draw.a.plot <-  function(p){
  plot(p)
  # title(age)
}

# Pure Top Down - Highest in OutStrength, Lowest in InStrength
loop.animate <- function(p) {
    map(p, draw.a.plot)
    # lapply(names(p), function(i) {draw.a.plot(i)})
}

gif_fun <- function(d, study, tname, sname){
  plts <- d$p
  names(plts) <- d$p
  saveGIF(loop.animate(plts)
          , interval = .5
          , movie.name = sprintf("%s/05-results/04-plots/02-q2/01-idio/03-vector-field-gifs/%s_%s_%s.gif", wd, study, tname, sname)
          , ani.width = 500
          , ani.height = 500
          )
}

q2_idio_gifs <- q2_idio_pred_p %>%
  select(-data) %>%
  group_by(study, t_name, s_name) %>%
  nest() %>%
  ungroup() %>%
  mutate(gif = pmap(list(data, study, t_name, s_name), gif_fun))
```


#### Tables  

```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/01-idio/01-models/%s", wd, file)
    load(path)
    get(ls()[ls() == obj])
    r <- VarCorr(m)$residual__$cor[,,"pdeltavalue"]["sdeltavalue",] %>% data.frame(.) %>% t() %>% as_tibble()
    rm(m)
    gc()
    return(r)
}


plan(multisession(workers = 16L))
q2_idio_cor <- tibble(file = sprintf("%s/05-results/02-q2/01-idio/01-models", wd) %>% list.files()) %>%
  separate(file, c("study", "SID", "t_name", "t_item", "s_name"), sep = "-", remove = F) %>%
  filter(t_name == t_item & !is.na(s_name)) %>%
  mutate(r = future_map2(
    file, "m"
    , possibly(loadRData, NA_real_)
    , .progress = T
    , .options = furrr_options(
      globals = c("wd")
      , packages = c("plyr", "tidyverse", "brms")
    )
    )) %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  select(-file)
closeAllConnections()
```


### Multilevel (Centered)  
#### Figures  
##### Forest Plots  
###### Fixed Effects with Credible Intervals  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/02-mlm/02-summary/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_fx <- tibble(file = sprintf("%s/05-results/02-q2/02-mlm/02-summary", wd) %>% list.files()) %>%
  mutate(fx = map2(file, "fx", loadRData)
         , rx = map2(file, "rx", loadRData)) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  left_join(expected)

q2_mlm_fx_fp <- q2_mlm_fx %>%
  select(-rx) %>%
  unnest(fx) %>%
  filter(t_name == t_item & study != "pairs") %>%
  filter(response %in% c("pdeltavalue", "sdeltavalue") & term == "p_value_c:s_value_c") %>%
  mutate(t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new))
         , s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
         , sig = ifelse(sign(conf.low) == sign(conf.high), "sig", "ns")
         , response = mapvalues(response, c("pdeltavalue", "sdeltavalue"), c("Change in Personality States", "Change in Situation Characteristics"))
         , study = factor(study, c("ipcs", "sherman", "horstmann"), c("Sample #1\n(Beck & Jackson)", "Sample #2\n(Sherman et al.)", "Sample #3\n(Horstmann et al.)")))

# total expected 13
# total unexpected 27

q2_mlm_fx_fp %>% 
  group_by(study, expy, sig) %>% 
  tally() %>%
  mutate(denom = ifelse(expy == "yes", 26, 52)
         , perc = n/denom*100) %>%
  select(-n) %>%
  pivot_wider(names_from = "expy", values_from = "perc") %>%
  arrange(rev(sig), study) %>%
  kable(., "html") %>%
  kable_classic()

q2_mlm_fx_fp %>% 
  group_by(study, expy, sig) %>% 
  summarize(estimate = mean(abs(estimate)))
  group_by(study) %>% 
  mutate(perc = sprintf("%i (%.2f%%)", n ,  n/sum(n)*100)) %>%
  select(-n) %>%
  pivot_wider(names_from = "expy", values_from = "perc") %>%
  kable(., "html") %>%
  kable_classic()

q2_mlm_fx_fp %>%
  ggplot(aes(x = t_name, y = estimate,shape = response)) + 
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high, group = response, color = sig)
                  , position = position_dodge(width = .9)
                  , width = .1) +
    # geom_errorbar(data = q2_mlm_fx_fp %>% filter(sig == "sig")
    #               , aes(ymin = conf.low, ymax = conf.high, group = response)
    #               , position = position_dodge(width = .9)
    #               , width = .1) +
    scale_y_continuous(limits = c(-.04, .04)
                       , breaks = seq(-.03, .03, .03)
                       , labels = c("-.03", "0", ".03")) +
    scale_color_manual(values = c("grey70", "black")) +
    geom_point(aes(shape = response, color = sig)
               , size = 1.5
               , position = position_dodge(width = .9)) +
    # geom_point(data = q2_mlm_fx_fp %>% filter(sig == "sig")
    #            , aes(shape = response)
    #               # , color = "white"
    #            , size = 1.5
    #            , position = position_dodge(width = .9)) +
    geom_point(data = q2_mlm_fx_fp %>% filter(expy == "yes")
               , aes(y = -0.03), shape = 8, color = "blue") +
    geom_hline(aes(yintercept = 0), linetype = "dashed") +
    labs(x = NULL, y = "Fixed Effect Attractor Estimate", shape = "") + 
    coord_flip() +
    facet_grid(study ~ s_name, scales = "free_y", space = "free") + 
    theme_classic() +
    guides(color = "none") +
    theme(axis.text = element_text(face = "bold", size = rel(1.1))
          , axis.title = element_text(face = "bold", size = rel(1.2))
          , strip.text.x = element_text(face = "bold", size = rel(1.1), color = "white")
          , strip.text.y = element_text(face = "bold", size = rel(1), color = "white")
          , strip.background = element_rect(fill = "black")
          , panel.background = element_rect(color = "black", size = 1)
          , legend.position = "bottom"
          )
ggsave(file = sprintf("%s/05-results/04-plots/02-q2/02-mlm/fx/traits.png", wd)
       , width = 10, height = 5)
```

```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/02-mlm/03-equil/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_fx <- tibble(file = sprintf("%s/05-results/02-q2/02-mlm/03-equil", wd) %>% list.files()) %>%
  mutate(equil = map2(file, "equil", loadRData)) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  left_join(expected)
```


##### Vector Fields  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/02-mlm/04-predictions/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_pred <- tibble(file = sprintf("%s/05-results/02-q2/02-mlm/04-predictions", wd) %>% list.files()) %>%
  mutate(fx_pred = map2(file, "fx_pred", possibly(loadRData, NA_real_))) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) 
```

```{r}
function(d, study, sname){
  sname2 <- mapvalues(sname, cat_names$name_old, cat_names$name_new, warn_missing = F)
  d %>% 
    # filter(p_value_c %in% seq(0,10,.4) & s_value_c %in% seq(0,10,.4)) %>%
    mutate(col = sqrt(Estimate_pdeltavalue^2 + Estimate_sdeltavalue^2)) %>%
    ggplot(aes(x = p_value, y = s_value, u = Estimate_pdeltavalue, v = Estimate_sdeltavalue)) + 
    # geom_point(data = df4, color = "white") +
    geom_quiver(aes(color = col))   +
    scale_color_gradient(low = "gray80", high = "black")+
    labs(x = "Personality State Level", y = sprintf("%s Level", sname2)) +
    # facet_wrap(~t_name) + 
    theme_classic() +
    theme(legend.position = "none",
          strip.background = element_rect(fill = "black"),
          strip.text = element_text(face = "bold", color = "white", size = rel(1.2)),
          axis.text = element_text(face = "bold", color = "black"),
          axis.title = element_text(face = "bold", size = rel(1.1)))
}

q2_mlm_pred_fx <- q2_mlm_pred %>%
  unnest(fx_pred) %>%
  filter(t_name == t_item) %>% 
  group_by(study, s_name) %>%
  nest() %>%
  ungroup()
```

#### Tables  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/02-mlm/01-models/%s", wd, file)
    load(path)
    get(ls()[ls() == obj])
    r <- VarCorr(m)$residual__$cor[,,"pdeltavalue"]["sdeltavalue",] %>% data.frame() %>% t() %>% as_tibble()
    rm(m)
    gc() 
    return(r)
}

plan(multisession(workers = 16L))
q2_mlm_cor <- tibble(file = sprintf("%s/05-results/02-q2/02-mlm/01-models", wd) %>% list.files()) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-", remove = F) %>%
  filter(t_name == t_item & !is.na(s_name)) %>%
  mutate(rdx = future_map2(
    file, "m"
    , possibly(loadRData, NA_real_)
    , .progress = T
    , .options = furrr_options(
      globals = c("wd")
      , packages = c("plyr", "tidyverse", "brms")
    )
    )) %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  select(-file)
closeAllConnections()
```

```{r}
q2_cors <- q2_mlm_cor %>%
  unnest(rdx) %>%
  mutate(SID = "Overall") %>%
  full_join(
    q2_idio_cor %>% 
      filter(!is.na(r)) %>% 
      unnest(r)
  ) %>%
  filter(study != "pairs") %>%
  mutate(s_name = str_remove_all(s_name, "[.]")) %>%
  mutate(
    s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
    , t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new))
    , study = factor(study, c("ipcs", "sherman", "horstmann"), c("Sample #1\n(Beck & Jackson)", "Sample #2\n(Sherman et al.)", "Sample #3\n(Horstmann et al.)"))
    , sig = ifelse(sign(`Q2.5`) == sign(Q97.5), "sig", "ns")
    ) %>%
  group_by(study, SID, t_name, s_name) %>% slice(1) %>%
  ungroup() %>%
  left_join(expected %>%
              mutate(s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
                     , t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new)))
    )
save(q2_cors, file = sprintf("%s/05-results/02-q2/q2-cors.RData", wd))
q2_cors %>%
  # filter(is.na(t_name))
  filter(!is.na(Estimate)) %>%
  ggplot(aes(y = t_name, x = Estimate, group = t_name)) +
    # geom_density_ridges(aes(group = t_name, fill = t_name)) +
    geom_density_ridges(aes(group = t_name), fill = "white", color = "white") +
    scale_x_continuous(limits = c(-1.05, 1.05), breaks = seq(-1,1,.5), labels = c("-1", "-.5", "0", ".5", "1")) +
    # geom_errorbar(data = q2_cors %>% filter(SID == "Overall" & sig == "sig")
    #               , aes(xmin = `Q2.5`, xmax = `Q97.5`)
    #               , position = "dodge"
    #               , width = .1) +
    # geom_point(data = q2_cors %>% filter(SID == "Overall" & sig == "sig")) +
    # geom_errorbar(data = q2_cors %>% filter(SID == "Overall" & sig == "ns")
    #               , aes(xmin = `Q2.5`, xmax = `Q97.5`)
    #               , position = "dodge"
    #               , color = "grey60"
    #               , width = .1) +
    # geom_point(data = q2_cors %>% filter(SID == "Overall" & sig == "ns")
    #            , color = "grey60") +
    geom_point(data = q2_cors %>% filter(expy == "yes")
               , aes(x = -1), shape = 8, color = "blue"
               , position = position_dodge(width = .8)) +
    geom_vline(aes(xintercept = 0), linetype = "dashed") +
    labs(y = NULL
         , x = "Residual (Bivariate) Correlation"
         ) + 
    # coord_flip() +
    facet_grid(study ~ s_name, scales = "free_y", space = "free") + 
    theme_classic() +
    theme(axis.text = element_text(face = "bold", size = rel(1.1))
          , axis.title = element_text(face = "bold", size = rel(1.2))
          , strip.text = element_text(face = "bold", size = rel(1), color = "white")
          , strip.background = element_rect(fill = "black")
          , panel.background = element_rect(color = "black", size = 1)
          , legend.position = "none")
  ggsave(file = sprintf("%s/05-results/04-plots/02-q2/cor-dist-exp.png", wd)
       , width = 10, height = 5)

```

```{r}
q2_cors %>% filter(SID == "Overall") %>% group_by(study, expy, sig) %>% tally() %>% 
  mutate(denom = ifelse(expy == "yes", 13, 27)
         , perc = n/denom*100)
```

### Multilevel (Uncentered)  
#### Figures  
##### Forest Plots  
###### Fixed Effects with Credible Intervals  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/03-mlm-uncentered/02-summary/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_fx <- tibble(file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/02-summary", wd) %>% list.files()) %>%
  mutate(fx = map2(file, "fx", loadRData)
         , rx = map2(file, "rx", loadRData)) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  left_join(expected)

q2_mlm_fx_fp <- q2_mlm_fx %>%
  select(-rx) %>%
  unnest(fx) %>%
  filter(t_name == t_item & study != "pairs") %>%
  filter(response %in% c("pdeltavalue", "sdeltavalue") & term == "p_value_c:s_value_c") %>%
  mutate(t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new))
         , s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
         , sig = ifelse(sign(conf.low) == sign(conf.high), "sig", "ns")
         , response = mapvalues(response, c("pdeltavalue", "sdeltavalue"), c("Change in Personality States", "Change in Situation Characteristics"))
         , study = factor(study, c("ipcs", "sherman", "horstmann"), c("Sample #1\n(Beck & Jackson)", "Sample #2\n(Sherman et al.)", "Sample #3\n(Horstmann et al.)")))

# total expected 13
# total unexpected 27

q2_mlm_fx_fp %>% 
  group_by(study, expy, sig) %>% 
  tally() %>%
  mutate(denom = ifelse(expy == "yes", 26, 52)
         , perc = n/denom*100) %>%
  select(-n) %>%
  pivot_wider(names_from = "expy", values_from = "perc") %>%
  arrange(rev(sig), study) %>%
  kable(., "html") %>%
  kable_classic()

q2_mlm_fx_fp %>% 
  group_by(study, expy, sig) %>% 
  summarize(estimate = mean(abs(estimate)))
  group_by(study) %>% 
  mutate(perc = sprintf("%i (%.2f%%)", n ,  n/sum(n)*100)) %>%
  select(-n) %>%
  pivot_wider(names_from = "expy", values_from = "perc") %>%
  kable(., "html") %>%
  kable_classic()

q2_mlm_fx_fp %>%
  ggplot(aes(x = t_name, y = estimate,shape = response)) + 
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high, group = response, color = sig)
                  , position = position_dodge(width = .9)
                  , width = .1) +
    # geom_errorbar(data = q2_mlm_fx_fp %>% filter(sig == "sig")
    #               , aes(ymin = conf.low, ymax = conf.high, group = response)
    #               , position = position_dodge(width = .9)
    #               , width = .1) +
    scale_y_continuous(limits = c(-.04, .04)
                       , breaks = seq(-.03, .03, .03)
                       , labels = c("-.03", "0", ".03")) +
    scale_color_manual(values = c("grey70", "black")) +
    geom_point(aes(shape = response, color = sig)
               , size = 1.5
               , position = position_dodge(width = .9)) +
    # geom_point(data = q2_mlm_fx_fp %>% filter(sig == "sig")
    #            , aes(shape = response)
    #               # , color = "white"
    #            , size = 1.5
    #            , position = position_dodge(width = .9)) +
    geom_point(data = q2_mlm_fx_fp %>% filter(expy == "yes")
               , aes(y = -0.03), shape = 8, color = "blue") +
    geom_hline(aes(yintercept = 0), linetype = "dashed") +
    labs(x = NULL, y = "Fixed Effect Attractor Estimate", shape = "") + 
    coord_flip() +
    facet_grid(study ~ s_name, scales = "free_y", space = "free") + 
    theme_classic() +
    guides(color = "none") +
    theme(axis.text = element_text(face = "bold", size = rel(1.1))
          , axis.title = element_text(face = "bold", size = rel(1.2))
          , strip.text.x = element_text(face = "bold", size = rel(1.1), color = "white")
          , strip.text.y = element_text(face = "bold", size = rel(1), color = "white")
          , strip.background = element_rect(fill = "black")
          , panel.background = element_rect(color = "black", size = 1)
          , legend.position = "bottom"
          )
ggsave(file = sprintf("%s/05-results/04-plots/02-q2/03-mlm-uncentered/fx/traits.png", wd)
       , width = 10, height = 5)
```

```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/03-mlm-uncentered/03-equil/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_fx <- tibble(file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/03-equil", wd) %>% list.files()) %>%
  mutate(equil = map2(file, "equil", loadRData)) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  left_join(expected)
```


##### Vector Fields  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/03-mlm-uncentered/04-predictions/%s", wd, file)
    load(path)
    get(ls()[grepl(obj, ls())])
}

q2_mlm_pred <- tibble(file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/04-predictions", wd) %>% list.files()) %>%
  mutate(fx_pred = map2(file, "fx_pred", possibly(loadRData, NA_real_))) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-") %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) 
```

```{r}
function(d, study, sname){
  sname2 <- mapvalues(sname, cat_names$name_old, cat_names$name_new, warn_missing = F)
  d %>% 
    # filter(p_value_c %in% seq(0,10,.4) & s_value_c %in% seq(0,10,.4)) %>%
    mutate(col = sqrt(Estimate_pdeltavalue^2 + Estimate_sdeltavalue^2)) %>%
    ggplot(aes(x = p_value, y = s_value, u = Estimate_pdeltavalue, v = Estimate_sdeltavalue)) + 
    # geom_point(data = df4, color = "white") +
    geom_quiver(aes(color = col))   +
    scale_color_gradient(low = "gray80", high = "black")+
    labs(x = "Personality State Level", y = sprintf("%s Level", sname2)) +
    # facet_wrap(~t_name) + 
    theme_classic() +
    theme(legend.position = "none",
          strip.background = element_rect(fill = "black"),
          strip.text = element_text(face = "bold", color = "white", size = rel(1.2)),
          axis.text = element_text(face = "bold", color = "black"),
          axis.title = element_text(face = "bold", size = rel(1.1)))
}

q2_mlm_pred_fx <- q2_mlm_pred %>%
  unnest(fx_pred) %>%
  filter(t_name == t_item) %>% 
  group_by(study, s_name) %>%
  nest() %>%
  ungroup()
```

#### Tables  
```{r}
loadRData <- function(file, obj){
    #loads an RData file, and returns it
    path <- sprintf("%s/05-results/02-q2/03-mlm-uncentered/01-models/%s", wd, file)
    load(path)
    get(ls()[ls() == obj])
    r <- VarCorr(m)$residual__$cor[,,"pdeltavalue"]["sdeltavalue",] %>% data.frame() %>% t() %>% as_tibble()
    rm(m)
    gc() 
    return(r)
}

plan(multisession(workers = 16L))
q2_mlm_cor <- tibble(file = sprintf("%s/05-results/02-q2/03-mlm-uncentered/01-models", wd) %>% list.files()) %>%
  separate(file, c("study", "t_name", "t_item", "s_name"), sep = "-", remove = F) %>%
  filter(t_name == t_item & !is.na(s_name)) %>%
  mutate(rdx = future_map2(
    file, "m"
    , possibly(loadRData, NA_real_)
    , .progress = T
    , .options = furrr_options(
      globals = c("wd")
      , packages = c("plyr", "tidyverse", "brms")
    )
    )) %>%
  mutate(s_name = str_remove(s_name, ".RData")) %>%
  mutate(s_name = ifelse(study == "ipcs" & s_name == "Sociability", "Sociality", s_name)) %>%
  select(-file)
closeAllConnections()
```

```{r}
q2_cors <- q2_mlm_cor %>%
  unnest(rdx) %>%
  mutate(SID = "Overall") %>%
  full_join(
    q2_idio_cor %>% 
      filter(!is.na(r)) %>% 
      unnest(r)
  ) %>%
  filter(study != "pairs") %>%
  mutate(s_name = str_remove_all(s_name, "[.]")) %>%
  mutate(
    s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
    , t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new))
    , study = factor(study, c("ipcs", "sherman", "horstmann"), c("Sample #1\n(Beck & Jackson)", "Sample #2\n(Sherman et al.)", "Sample #3\n(Horstmann et al.)"))
    , sig = ifelse(sign(`Q2.5`) == sign(Q97.5), "sig", "ns")
    ) %>%
  group_by(study, SID, t_name, s_name) %>% slice(1) %>%
  ungroup() %>%
  left_join(expected %>%
              mutate(s_name = factor(s_name, cat_names$name_old, cat_names$name_new)
                     , t_name = factor(t_name, rev(cat_names$name_old), rev(cat_names$name_new)))
    )
save(q2_cors, file = sprintf("%s/05-results/03-mlm-uncentered/q2-cors.RData", wd))
q2_cors %>%
  # filter(is.na(t_name))
  filter(!is.na(Estimate)) %>%
  ggplot(aes(y = t_name, x = Estimate, group = t_name)) +
    # geom_density_ridges(aes(group = t_name, fill = t_name)) +
    geom_density_ridges(aes(group = t_name), fill = "white", color = "white") +
    scale_x_continuous(limits = c(-1.05, 1.05), breaks = seq(-1,1,.5), labels = c("-1", "-.5", "0", ".5", "1")) +
    # geom_errorbar(data = q2_cors %>% filter(SID == "Overall" & sig == "sig")
    #               , aes(xmin = `Q2.5`, xmax = `Q97.5`)
    #               , position = "dodge"
    #               , width = .1) +
    # geom_point(data = q2_cors %>% filter(SID == "Overall" & sig == "sig")) +
    # geom_errorbar(data = q2_cors %>% filter(SID == "Overall" & sig == "ns")
    #               , aes(xmin = `Q2.5`, xmax = `Q97.5`)
    #               , position = "dodge"
    #               , color = "grey60"
    #               , width = .1) +
    # geom_point(data = q2_cors %>% filter(SID == "Overall" & sig == "ns")
    #            , color = "grey60") +
    geom_point(data = q2_cors %>% filter(expy == "yes")
               , aes(x = -1), shape = 8, color = "blue"
               , position = position_dodge(width = .8)) +
    geom_vline(aes(xintercept = 0), linetype = "dashed") +
    labs(y = NULL
         , x = "Residual (Bivariate) Correlation"
         ) + 
    # coord_flip() +
    facet_grid(study ~ s_name, scales = "free_y", space = "free") + 
    theme_classic() +
    theme(axis.text = element_text(face = "bold", size = rel(1.1))
          , axis.title = element_text(face = "bold", size = rel(1.2))
          , strip.text = element_text(face = "bold", size = rel(1), color = "white")
          , strip.background = element_rect(fill = "black")
          , panel.background = element_rect(color = "black", size = 1)
          , legend.position = "none")
  ggsave(file = sprintf("%s/05-results/04-plots/02-q2/cor-dist-exp.png", wd)
       , width = 10, height = 5)

```

```{r}
q2_cors %>% filter(SID == "Overall") %>% group_by(study, expy, sig) %>% tally() %>% 
  mutate(denom = ifelse(expy == "yes", 13, 27)
         , perc = n/denom*100)
```
